Adding explicit feign interceptor config to Idam
