+++ src/main/java/uk/gov/hmcts/reform/userprofileapi/domain/feign/IdamFeignClient.java
@@ -10,8 +10,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
+import uk.gov.hmcts.reform.userprofileapi.config.FeignInterceptorConfiguration;
 
-@FeignClient(name = "IdamFeignClient", url = "${auth.idam.client.baseUrl}")
+@FeignClient(name = "IdamFeignClient", url = "${auth.idam.client.baseUrl}", configuration = FeignInterceptorConfiguration.class)
 public interface IdamFeignClient {
 
     @RequestMapping(method = RequestMethod.POST, value = "/user/registration")
